refactor(svelte2tsx): コンポーネント名 / runes / script / generics 解析を nodes/ へ
- `nodes/component_name.rs`: `classNameFromFilename` と scule のケース変換移植
- `nodes/runes_detection.rs`: runes / await のテンプレート走査（相互再帰群なので 1 ファイル）
- `nodes/scripts.rs`: import 抽出とトップレベル await 検出（`nodes/Scripts.ts` 相当）
- `nodes/generics.rs`: `generics` 属性の解析と `$$ComponentProps` 巻き上げ可否判定

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
