+++ crates/rsvelte_core/src/svelte2tsx/nodes/mod.rs
@@ -1,0 +1,6 @@
+//! Per-node analyses and emitters, mirroring `svelte2tsx/src/svelte2tsx/nodes/`.
+
+pub mod component_name;
+pub mod generics;
+pub mod runes_detection;
+pub mod scripts;
