+++ crates/rsvelte_core/src/svelte2tsx/svelte2tsx.rs
@@ -10,12 +10,21 @@ use crate::compiler::phases::phase1_parse::{self, ParseOptions};
 
 use super::helpers::rewrite_external_imports::rewrite_external_specifiers_in_text;
 use super::magic_string::{GenerateMapOptions, MagicString};
+use super::nodes::component_name::derive_component_name;
+use super::nodes::generics::{
+    compact_generic_params, extract_generics_from_script_tag, split_generic_param_names,
+    type_text_references_any, type_text_typeof_references_local_value,
+};
+use super::nodes::runes_detection::{
+    detect_await_in_template, detect_rune_global_in_template, detect_runes_mode,
+};
+use super::nodes::scripts::{
+    detect_top_level_await, find_instance_imports, find_script_close_tag_start,
+};
 use super::script::{ComponentEvents, ExportedNames};
 use super::template;
 use super::utils::htmlxparser::{blank_style_content, blank_style_tags, remove_orphan_scripts};
-use super::utils::lexical::{
-    contains_word, is_ident_char, lexical_identifiers, lexical_identifiers_in_expressions,
-};
+use super::utils::lexical::{lexical_identifiers, lexical_identifiers_in_expressions};
 use super::validation::{validate_debug_tag_arguments, validate_meta_element_placement};
 
 pub use super::interfaces::{
@@ -2362,137 +2371,6 @@ fn extract_component_documentation(fragment: &crate::ast::template::Fragment) ->
     None
 }
 
-/// Find the start of `</script>` tag by scanning backwards from the script end position.
-fn find_script_close_tag_start(source: &str, script_end: u32) -> u32 {
-    let bytes = source.as_bytes();
-    let end = script_end as usize;
-    let needle = b"</script>";
-    let needle_len = needle.len();
-
-    if end < needle_len {
-        return script_end;
-    }
-
-    let mut i = end;
-    while i >= needle_len {
-        i -= 1;
-        if i + needle_len <= end
-            && bytes[i..i + needle_len]
-                .iter()
-                .zip(needle.iter())
-                .all(|(a, b)| a.to_ascii_lowercase() == *b)
-        {
-            return i as u32;
-        }
-    }
-
-    script_end
-}
-
-/// Find top-level import declarations in an instance script.
-///
-/// Returns a sorted list of (start, end) positions relative to the script
-/// content (i.e., relative to `script.content_offset`).
-/// Returns `(comments_start, import_start, import_end)` for each top-level
-/// import in `script`. `comments_start <= import_start` — the leading comment
-/// span lets the caller hoist JSDoc / line comments alongside their import,
-/// matching the JS reference's `moveNode` per-comment moves.
-fn find_instance_imports(
-    script: &crate::ast::template::Script,
-    source: &str,
-) -> Vec<(u32, u32, u32)> {
-    use oxc_allocator::Allocator;
-    use oxc_ast::ast as oxc;
-    use oxc_parser::Parser as OxcParser;
-    use oxc_span::SourceType;
-
-    let content_start = script.content_offset as usize;
-    let script_source = slice_src(source, script.start as usize, script.end as usize);
-    let close_tag_offset = script_source
-        .rfind("</script>")
-        .or_else(|| script_source.rfind("</Script>"))
-        .unwrap_or(script_source.len());
-    let content_end = script.start as usize + close_tag_offset;
-    let raw_content = &source[content_start..content_end];
-
-    // Fast path: an `import` substring is required for any import
-    // declaration to exist. Skip the OXC parse entirely for the majority
-    // of scripts that have no imports.
-    if !contains_word(raw_content.as_bytes(), b"import") {
-        return Vec::new();
-    }
-
-    let allocator = Allocator::default();
-    // Always use TypeScript source type for import detection.
-    // TypeScript is a superset of JavaScript, so TS parsing handles
-    // both `import type` (TS syntax) and regular imports correctly,
-    // even when the script doesn't have `lang="ts"`.
-    let source_type = SourceType::ts();
-    let parser = OxcParser::new(&allocator, raw_content, source_type);
-    let result = parser.parse();
-
-    // Comment spans (start, end) discovered by the parser, sorted by end. Used
-    // to compute each import's leading-comment region the way TS
-    // `getLeadingCommentRanges(node.getFullText())` does — including a TRAILING
-    // line comment on the PREVIOUS statement's line (it is leading trivia of the
-    // following import and moves up with it). The parser already tokenised
-    // strings/regex correctly, so `// …` inside a string is never misread.
-    let comment_spans: Vec<(u32, u32)> = result
-        .program
-        .comments
-        .iter()
-        .map(|c| (c.span.start, c.span.end))
-        .collect();
-
-    let mut imports = Vec::new();
-    let bytes = raw_content.as_bytes();
-    for stmt in result.program.body.iter() {
-        if let oxc::Statement::ImportDeclaration(import) = stmt {
-            // All import declarations (including side-effect imports like `import ''`)
-            // should be lifted. The parser only creates ImportDeclaration nodes for
-            // valid `import` statements with a source clause.
-            let start = import.span.start;
-            let end = import.span.end;
-
-            // Walk backwards over leading trivia, pulling in every comment whose
-            // end is reachable from the current start via whitespace only, and
-            // stopping at the first non-comment code (the previous token). This
-            // mirrors `getLeadingCommentRanges` and pulls a trailing line comment
-            // (`import …; // TODO`) into the FOLLOWING import's leading region.
-            let new_start = scan_back_leading_comments(bytes, start as usize, &comment_spans);
-
-            imports.push((new_start, start, end));
-        }
-    }
-    imports.sort_by_key(|&(s, _, _)| s);
-    imports
-}
-
-/// Walk backwards from `pos` over leading trivia (whitespace + comments),
-/// returning the start of the earliest comment reachable via whitespace-only
-/// gaps. Stops at the first non-comment code. `comment_spans` are parser-
-/// discovered `(start, end)` pairs (so strings/regex never produce false `//`).
-fn scan_back_leading_comments(bytes: &[u8], pos: usize, comment_spans: &[(u32, u32)]) -> u32 {
-    let mut cstart = pos as u32;
-    loop {
-        // Skip whitespace backward.
-        let mut p = cstart as usize;
-        while p > 0 && matches!(bytes[p - 1], b' ' | b'\t' | b'\n' | b'\r') {
-            p -= 1;
-        }
-        // A comment ending exactly at `p` is leading trivia of this import.
-        if let Some(&(cs, _)) = comment_spans.iter().find(|&&(_, ce)| ce as usize == p) {
-            if cs >= cstart {
-                break;
-            }
-            cstart = cs;
-        } else {
-            break;
-        }
-    }
-    cstart
-}
-
 /// Decide whether a top-level `{#snippet}` block is module-hoistable.
 ///
 /// A snippet is module-hoistable when its body's free variables resolve only
@@ -2615,344 +2493,6 @@ fn is_snippet_module_hoistable(
     true
 }
 
-/// Return true if `type_text` mentions any of `names` as a whole identifier
-/// (i.e. surrounded by non-identifier characters on both sides).
-///
-/// Used to detect when a `$$ComponentProps` body references a type/interface
-/// or value declared at the top level of the instance script — in which case
-/// the synthesised `;type $$ComponentProps = ...;` cannot be hoisted above
-/// `function $$render()`.
-fn type_text_references_any(type_text: &str, names: &std::collections::HashSet<String>) -> bool {
-    if names.is_empty() {
-        return false;
-    }
-    let bytes = type_text.as_bytes();
-    for name in names.iter() {
-        if name.is_empty() {
-            continue;
-        }
-        let nbytes = name.as_bytes();
-        let mut i = 0usize;
-        while i + nbytes.len() <= bytes.len() {
-            if &bytes[i..i + nbytes.len()] == nbytes {
-                let before_ok = i == 0 || !is_ident_char(bytes[i - 1]);
-                let after_idx = i + nbytes.len();
-                let after_ok = after_idx == bytes.len() || !is_ident_char(bytes[after_idx]);
-                if before_ok && after_ok {
-                    return true;
-                }
-            }
-            i += 1;
-        }
-    }
-    false
-}
-
-/// Return true if `type_text` contains a `typeof X` value-query whose root
-/// identifier `X` is a value **declared in the instance script** (and not an
-/// import).
-///
-/// This mirrors upstream's `HoistableInterfaces` hoistability test for the
-/// synthesised `$$ComponentProps` alias: a `typeof X` adds `X` to the props
-/// interface's `value_deps`, and the alias can only be hoisted above
-/// `function $$render()` when every value dep is an *allowed reference*
-/// (`isAllowedReference`) — i.e. NOT a locally-declared instance value. A
-/// `typeof` of an **imported** binding (e.g. `ComponentProps<typeof Button>`
-/// where `Button` is `import`ed) stays hoistable, because the import lives at
-/// module scope above `$$render`. The previous heuristic forced *every*
-/// `typeof` inside `$$render`, which wrongly nested the alias for the very
-/// common imported-component case.
-fn type_text_typeof_references_local_value(
-    type_text: &str,
-    instance_value_names: &std::collections::HashSet<String>,
-    instance_import_names: &std::collections::HashSet<String>,
-    module_import_names: &std::collections::HashSet<String>,
-) -> bool {
-    let bytes = type_text.as_bytes();
-    let kw = b"typeof";
-    let mut i = 0usize;
-    while i + kw.len() <= bytes.len() {
-        if &bytes[i..i + kw.len()] == kw {
-            let before_ok = i == 0 || !is_ident_char(bytes[i - 1]);
-            let mut j = i + kw.len();
-            // `typeof` must be followed by whitespace (a value query), not be a
-            // prefix of a longer identifier like `typeofX`.
-            let has_ws = j < bytes.len() && bytes[j].is_ascii_whitespace();
-            if before_ok && has_ws {
-                while j < bytes.len() && bytes[j].is_ascii_whitespace() {
-                    j += 1;
-                }
-                // Capture the root identifier (stop at `.` / non-ident), e.g.
-                // `foo.bar.baz` -> `foo`.
-                let start = j;
-                while j < bytes.len() && is_ident_char(bytes[j]) {
-                    j += 1;
-                }
-                if j > start {
-                    let root = &type_text[start..j];
-                    let is_import =
-                        instance_import_names.contains(root) || module_import_names.contains(root);
-                    if !is_import && instance_value_names.contains(root) {
-                        return true;
-                    }
-                }
-                i = j;
-                continue;
-            }
-        }
-        i += 1;
-    }
-    false
-}
-
-/// Split a generics string like "T extends Record<string, any>, U" into
-/// just the type parameter names: ["T", "U"].
-/// Handles nested angle brackets and commas inside constraints.
-fn split_generic_param_names(generics: &str) -> Vec<String> {
-    let mut names = Vec::new();
-    let mut depth = 0; // angle bracket depth
-    let mut current_start = 0;
-
-    for (i, ch) in generics.char_indices() {
-        match ch {
-            '<' => depth += 1,
-            '>' if depth > 0 => depth -= 1,
-            ',' if depth == 0 => {
-                let param = generics[current_start..i].trim();
-                names.push(extract_param_name(param));
-                current_start = i + 1;
-            }
-            _ => {}
-        }
-    }
-    // Handle the last parameter
-    let param = generics[current_start..].trim();
-    if !param.is_empty() {
-        names.push(extract_param_name(param));
-    }
-    names
-}
-
-/// Compact a generics string by stripping leading spaces from each top-level parameter.
-/// "A, B extends keyof A, C extends boolean" → "A,B extends keyof A,C extends boolean"
-fn compact_generic_params(generics: &str) -> String {
-    let mut result = String::new();
-    let mut depth = 0;
-    let mut after_comma = false;
-
-    for ch in generics.chars() {
-        match ch {
-            '<' => {
-                depth += 1;
-                result.push(ch);
-            }
-            '>' => {
-                if depth > 0 {
-                    depth -= 1;
-                }
-                result.push(ch);
-            }
-            ',' if depth == 0 => {
-                result.push(',');
-                after_comma = true;
-            }
-            ' ' | '\t' if after_comma => {
-                // Skip leading whitespace after comma at top level
-                continue;
-            }
-            _ => {
-                after_comma = false;
-                result.push(ch);
-            }
-        }
-    }
-    result
-}
-
-/// Extract the type parameter name from a parameter declaration,
-/// handling the `const` modifier (e.g., `const T extends ...` → `T`).
-fn extract_param_name(param: &str) -> String {
-    let mut words = param.split_whitespace();
-    let first = words.next().unwrap_or("");
-    if first == "const" {
-        // Skip `const` modifier, take the next word
-        words.next().unwrap_or(first).to_string()
-    } else {
-        first.to_string()
-    }
-}
-
-/// Detect whether a script content contains top-level `await` expressions.
-///
-/// Uses OXC to parse the content as a module (which allows top-level await)
-/// and checks for AwaitExpression at the top level of the program body.
-fn detect_top_level_await(content: &str) -> bool {
-    use oxc_allocator::Allocator;
-    use oxc_ast::ast as oxc;
-    use oxc_parser::Parser as OxcParser;
-    use oxc_span::SourceType;
-
-    // Fast path: an `await` substring is required for any top-level await
-    // to exist. Skip the OXC parse entirely when the keyword is absent.
-    if !contains_word(content.as_bytes(), b"await") {
-        return false;
-    }
-
-    let allocator = Allocator::default();
-    let source_type = SourceType::ts().with_module(true);
-    let parser = OxcParser::new(&allocator, content, source_type);
-    let result = parser.parse();
-
-    // Mirror upstream `processInstanceScriptContent.ts` which sets
-    // `hasTopLevelAwait = true` whenever an AwaitExpression is visited at the
-    // root scope (i.e. not inside any Block / FunctionLike node).
-    //
-    // We do not have the upstream's full AST-walker machinery, but we can
-    // replicate the effect for the cases that actually occur in Svelte
-    // components:
-    //
-    //   • `VariableDeclaration` at module top-level whose initialiser
-    //     *contains* an AwaitExpression (e.g. `let x = $derived(await f())`
-    //     or `const user = await getUser()`).
-    //   • `ExpressionStatement` at module top-level whose expression
-    //     *contains* an AwaitExpression (e.g. `y = await promise`).
-    //
-    // For both, we use a deep recursive scan that stops at function
-    // boundaries (`FunctionExpression` / `ArrowFunctionExpression`) — those
-    // introduce a new scope and their inner `await` is NOT top-level.
-    for stmt in result.program.body.iter() {
-        match stmt {
-            oxc::Statement::VariableDeclaration(decl) => {
-                for declarator in decl.declarations.iter() {
-                    if let Some(ref init) = declarator.init
-                        && expr_contains_await_deep(init)
-                    {
-                        return true;
-                    }
-                }
-            }
-            oxc::Statement::ExpressionStatement(expr)
-                if expr_contains_await_deep(&expr.expression) =>
-            {
-                return true;
-            }
-            _ => {}
-        }
-    }
-    false
-}
-
-/// Deep check: returns `true` if `expr` is, or transitively contains (outside
-/// any function boundary), an `AwaitExpression`.
-///
-/// Mirrors the upstream TypeScript walker's `scope === rootScope` predicate:
-/// we recurse into all expression sub-nodes but stop when entering a
-/// `FunctionExpression` or `ArrowFunctionExpression` (a new async scope
-/// whose internal `await` is not "top-level" for the Svelte component).
-///
-/// Reference: `processInstanceScriptContent.ts` lines 246-349
-///   `if (ts.isBlock(node) || ts.isFunctionLike(node)) pushScope();`
-///   `if (isSvelte5Plus && ts.isAwaitExpression(node) && scope === rootScope)`
-fn expr_contains_await_deep(expr: &oxc_ast::ast::Expression) -> bool {
-    use oxc_ast::ast::{Argument, Expression as E};
-
-    match expr {
-        // Base case: this expression is an await.
-        E::AwaitExpression(_) => true,
-
-        // Function boundaries: do NOT recurse — their inner awaits are in a
-        // new scope and are not "top-level" from the component's perspective.
-        E::ArrowFunctionExpression(_) | E::FunctionExpression(_) => false,
-
-        // Parenthesised expression: transparent wrapper.
-        E::ParenthesizedExpression(p) => expr_contains_await_deep(&p.expression),
-
-        // Assignment: `x = await y` or `x = f(await y)` — check the RHS.
-        // (LHS is a pattern/identifier and cannot contain await directly.)
-        E::AssignmentExpression(a) => expr_contains_await_deep(&a.right),
-
-        // Binary / logical: check both sides.
-        E::BinaryExpression(b) => {
-            expr_contains_await_deep(&b.left) || expr_contains_await_deep(&b.right)
-        }
-        E::LogicalExpression(l) => {
-            expr_contains_await_deep(&l.left) || expr_contains_await_deep(&l.right)
-        }
-
-        // Conditional: test ? consequent : alternate.
-        E::ConditionalExpression(c) => {
-            expr_contains_await_deep(&c.test)
-                || expr_contains_await_deep(&c.consequent)
-                || expr_contains_await_deep(&c.alternate)
-        }
-
-        // Unary / yield: single argument.
-        E::UnaryExpression(u) => expr_contains_await_deep(&u.argument),
-        E::YieldExpression(y) => y
-            .argument
-            .as_ref()
-            .is_some_and(|a| expr_contains_await_deep(a)),
-
-        // Sequence: any expression in the list.
-        E::SequenceExpression(s) => s.expressions.iter().any(expr_contains_await_deep),
-
-        // Call expression: callee + arguments.  This is the key case for
-        // `$derived(await x)` — the callee is `$derived` and the argument
-        // is an AwaitExpression.
-        //
-        // `Argument` inherits `Expression` variants via `@inherit Expression`;
-        // `to_expression()` panics for `SpreadElement` (handled first in the
-        // match), and returns the inner `&Expression` for all other variants.
-        E::CallExpression(call) => {
-            expr_contains_await_deep(&call.callee)
-                || call.arguments.iter().any(|arg| match arg {
-                    Argument::SpreadElement(sp) => expr_contains_await_deep(&sp.argument),
-                    _ => expr_contains_await_deep(arg.to_expression()),
-                })
-        }
-
-        // `new Foo(await x)`.
-        E::NewExpression(n) => n.arguments.iter().any(|arg| match arg {
-            Argument::SpreadElement(sp) => expr_contains_await_deep(&sp.argument),
-            _ => expr_contains_await_deep(arg.to_expression()),
-        }),
-
-        // Member expressions: `obj[await key]` or `obj.prop`.
-        E::ComputedMemberExpression(c) => {
-            expr_contains_await_deep(&c.object) || expr_contains_await_deep(&c.expression)
-        }
-        E::StaticMemberExpression(s) => expr_contains_await_deep(&s.object),
-        E::PrivateFieldExpression(p) => expr_contains_await_deep(&p.object),
-
-        // Template literals: `${await x}`.
-        E::TemplateLiteral(tl) => tl.expressions.iter().any(expr_contains_await_deep),
-        E::TaggedTemplateExpression(tt) => {
-            expr_contains_await_deep(&tt.tag)
-                || tt.quasi.expressions.iter().any(expr_contains_await_deep)
-        }
-
-        // Object / array literals: `$derived({ value: await x })`, `[await x]`.
-        E::ObjectExpression(o) => o.properties.iter().any(|p| match p {
-            oxc_ast::ast::ObjectPropertyKind::ObjectProperty(prop) => {
-                expr_contains_await_deep(&prop.value)
-            }
-            oxc_ast::ast::ObjectPropertyKind::SpreadProperty(sp) => {
-                expr_contains_await_deep(&sp.argument)
-            }
-        }),
-        E::ArrayExpression(a) => a.elements.iter().any(|el| match el {
-            oxc_ast::ast::ArrayExpressionElement::SpreadElement(sp) => {
-                expr_contains_await_deep(&sp.argument)
-            }
-            oxc_ast::ast::ArrayExpressionElement::Elision(_) => false,
-            other => other.as_expression().is_some_and(expr_contains_await_deep),
-        }),
-
-        // Everything else (literals, identifiers, `this`, …) cannot contain await.
-        _ => false,
-    }
-}
-
 /// Count whitespace between tag name and first attribute in source.
 fn count_tag_to_attr_spaces_in_source(tag_name: &str, el_start: u32, source: &str) -> usize {
     let name_end = el_start as usize + 1 + tag_name.len(); // +1 for '<'
@@ -2969,281 +2509,6 @@ fn count_tag_to_attr_spaces_in_source(tag_name: &str, el_start: u32, source: &st
         }
     }
     count
-}
-
-/// Extract the `generics` attribute value from a script tag text.
-fn extract_generics_from_script_tag(tag_text: &str) -> Option<String> {
-    if let Some(pos) = tag_text.find("generics=") {
-        let after = &tag_text[pos + 9..];
-        let trimmed = after.trim_start();
-        if let Some(quote_char) = trimmed.chars().next() {
-            if quote_char == '"' || quote_char == '\'' {
-                let content = &trimmed[1..];
-                if let Some(end) = content.find(quote_char) {
-                    return Some(content[..end].to_string());
-                }
-            } else {
-                // Unquoted value: take until whitespace or `>`
-                let end = trimmed
-                    .find(|c: char| c.is_whitespace() || c == '>')
-                    .unwrap_or(trimmed.len());
-                if end > 0 {
-                    return Some(trimmed[..end].to_string());
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Derive a safe component name from the filename.
-///
-/// Converts "App.svelte" -> "App", "my-component.svelte" -> "My_component",
-/// handles path separators and special characters.
-///
-/// Port of `classNameFromFilename` from
-/// `submodules/language-tools/packages/svelte2tsx/src/svelte2tsx/addComponentExport.ts`.
-///
-/// Algorithm:
-/// 1. Take the final path segment (after the last `/`), then everything before the
-///    first `.` — this is `withoutExtensions`.
-/// 2. Keep only `[A-Za-z_\d-]` characters — `withoutInvalidCharacters`.
-/// 3. Find the index of the first ASCII letter (`firstValidCharIdx`).
-/// 4. `withoutLeadingInvalidCharacters = withoutInvalidCharacters.substr(firstValidCharIdx)`.
-///    JS `substr(-1)` (when no letter is found, idx = -1) returns the **last character**
-///    of the string.
-/// 5. `inPascalCase = scule_pascal_case(withoutLeadingInvalidCharacters)`.
-/// 6. If no letter was found (`firstValidCharIdx == -1`), prepend `"A"`.
-fn derive_component_name(filename: &str) -> String {
-    // Step 1: basename up to first dot  (mirrors `path.parse(filename).name?.split('.')[0]`)
-    let basename = filename.rsplit('/').next().unwrap_or(filename);
-    let basename = basename.rsplit('\\').next().unwrap_or(basename);
-    let without_extensions = basename.split('.').next().unwrap_or("");
-
-    // Step 2: keep only [A-Za-z_\d-]
-    let without_invalid: String = without_extensions
-        .chars()
-        .filter(|c| c.is_ascii_alphanumeric() || *c == '_' || *c == '-')
-        .collect();
-
-    // Step 3: find first ASCII letter
-    let first_valid_char_idx: Option<usize> = without_invalid
-        .chars()
-        .position(|c| c.is_ascii_alphabetic());
-
-    // Step 4: JS substr(firstValidCharIdx)
-    //   - When idx == -1 (no letter), JS substr(-1) returns the LAST character.
-    //   - When idx >= 0, take from that index onward.
-    let without_leading: &str = match first_valid_char_idx {
-        Some(idx) => {
-            // idx is a char-position; since all chars are ASCII, byte == char index.
-            &without_invalid[idx..]
-        }
-        None => {
-            // No ASCII letter: mimic JS substr(-1) → last character of the string.
-            // If the string is empty, this yields "" (empty slice).
-            if without_invalid.is_empty() {
-                ""
-            } else {
-                let last_char_byte = without_invalid
-                    .char_indices()
-                    .last()
-                    .map(|(i, _)| i)
-                    .unwrap_or(0);
-                &without_invalid[last_char_byte..]
-            }
-        }
-    };
-
-    // Step 5: apply scule pascalCase
-    let in_pascal_case = scule_pascal_case(without_leading);
-
-    // Step 6: prepend "A" when no letter was present
-    if first_valid_char_idx.is_none() {
-        format!("A{}", in_pascal_case)
-    } else {
-        in_pascal_case
-    }
-}
-
-/// Port of scule's `pascalCase` (no-normalize variant used by svelte2tsx).
-///
-/// `pascalCase(str) = splitByCase(str).map(upperFirst).join("")`
-///
-/// Reference: `node_modules/scule/dist/index.mjs` (used by svelte2tsx).
-fn scule_pascal_case(s: &str) -> String {
-    split_by_case(s)
-        .into_iter()
-        .map(|part| upper_first(&part))
-        .collect()
-}
-
-/// Uppercase only the first character of a string (scule `upperFirst`).
-fn upper_first(s: &str) -> String {
-    let mut chars = s.chars();
-    match chars.next() {
-        None => String::new(),
-        Some(first) => {
-            let mut result = String::with_capacity(s.len());
-            for c in first.to_uppercase() {
-                result.push(c);
-            }
-            result.extend(chars);
-            result
-        }
-    }
-}
-
-/// Port of scule's `splitByCase` with default splitters `["-", "_", "/", "."]`.
-///
-/// Three-state `isUppercase`:
-///   - digit  → `None`      (never triggers a split)
-///   - upper  → `Some(true)`
-///   - lower  → `Some(false)`
-///
-/// Splits occur:
-///   - On a splitter character (push current buffer, reset).
-///   - On a lower→UPPER transition (camelCase boundary).
-///   - On a UPPER→lower transition when buffer length > 1 (e.g. "ABCWidget" → ["ABC","Widget"]).
-///
-/// `previousSplitter` starts as `None` (not-false), so the transition checks are
-/// skipped for the very first character.
-fn split_by_case(s: &str) -> Vec<String> {
-    const SPLITTERS: [char; 4] = ['-', '_', '/', '.'];
-
-    let mut parts: Vec<String> = Vec::new();
-    let mut buff = String::new();
-    // Three-state: None = unset (first char), Some(true) = was splitter, Some(false) = not splitter
-    let mut previous_splitter: Option<bool> = None;
-    let mut previous_upper: Option<bool> = None; // None = digit/unset, Some(true/false)
-
-    for ch in s.chars() {
-        let is_splitter = SPLITTERS.contains(&ch);
-
-        if is_splitter {
-            parts.push(buff.clone());
-            buff.clear();
-            previous_upper = None;
-            previous_splitter = Some(true);
-            continue;
-        }
-
-        // isUppercase: digit → None, else compare with lowercase
-        let is_upper: Option<bool> = if ch.is_ascii_digit() {
-            None
-        } else if ch.is_uppercase() {
-            Some(true)
-        } else {
-            Some(false)
-        };
-
-        // Transition checks only when previousSplitter === false (not a splitter and not unset)
-        if previous_splitter == Some(false) {
-            // lower → UPPER: start a new part
-            if previous_upper == Some(false) && is_upper == Some(true) {
-                parts.push(buff.clone());
-                buff.clear();
-                buff.push(ch);
-                previous_upper = is_upper;
-                previous_splitter = Some(false);
-                continue;
-            }
-            // UPPER → lower when buff.len() > 1: split off all-but-last char of buffer
-            if previous_upper == Some(true) && is_upper == Some(false) && buff.len() > 1 {
-                let last_char = buff.chars().last().unwrap();
-                let split_point = buff.len() - last_char.len_utf8();
-                parts.push(buff[..split_point].to_string());
-                buff = format!("{}{}", last_char, ch);
-                previous_upper = is_upper;
-                previous_splitter = Some(false);
-                continue;
-            }
-        }
-
-        buff.push(ch);
-        previous_upper = is_upper;
-        previous_splitter = Some(false);
-    }
-
-    parts.push(buff);
-    parts
-}
-
-#[cfg(test)]
-mod scule_tests {
-    use super::*;
-
-    #[test]
-    fn test_split_by_case_basics() {
-        assert_eq!(split_by_case("my-component"), vec!["my", "component"]);
-        // "ABCWidget": UPPER→lower fires on 'i' after buff="ABCW" → ["ABC","Widget"]
-        assert_eq!(split_by_case("ABCWidget"), vec!["ABC", "Widget"]);
-        // "XMLHttp": UPPER→lower fires on 't' after buff="XMLH" → ["XML","Http"]
-        assert_eq!(split_by_case("XMLHttp"), vec!["XML", "Http"]);
-        assert_eq!(split_by_case("a1b2"), vec!["a1b2"]);
-    }
-}
-
-// NOTE on runes detection: svelte2tsx deliberately uses its OWN runes heuristic
-// (the `detect_*`/`ExportedNames::is_runes_mode` machinery) rather than the
-// compiler's authoritative `ComponentAnalysis::runes` flag. The two genuinely
-// DIVERGE: the compiler treats `$host` / `$inspect` / `$bindable` (and certain
-// shadowing/scope cases) as runes — semantically correct — but official
-// `svelte2tsx`'s `hasRunesGlobals` only counts `$state` / `$derived` / `$effect`
-// (plus `$props` / explicit / top-level await). Since this port targets
-// byte-parity with official `svelte2tsx`, it must mirror svelte2tsx's narrower
-// definition; wiring in the compiler flag was measured to REGRESS the corpus
-// (it over-detects runes for ~24 `$host`-only / shadowed-derived components).
-
-/// Detect whether the component uses Svelte 5 runes mode.
-///
-/// Checks for the presence of `$props()`, `$state()`, `$derived()`, etc. in script content,
-/// or `runes: true` in `<svelte:options>`.
-fn detect_runes_mode(ast: &Root) -> bool {
-    // Check svelte:options for explicit runes setting
-    if let Some(ref options) = ast.options
-        && let Some(runes) = options.runes
-    {
-        return runes;
-    }
-
-    // Don't default to runes mode; let process_instance_script detect rune usage
-    false
-}
-
-/// Detect `await` expressions inside template expression tags, e.g. `{await t}`.
-///
-/// This walks the template fragment AST looking for `ExpressionTag` nodes whose
-/// expression is (or begins with) an `AwaitExpression`. Await-in-template forces
-/// runes mode — async template expressions are Svelte 5 runes-only.
-///
-/// NOTE: `{#await ...}` block syntax is NOT detected here — only bare `await`
-/// inside `{...}` expression tags counts.
-///
-/// Reference: language-tools/packages/svelte2tsx/src/svelte2tsx/nodes/ExportedNames.ts
-///   `isRunes = true when component has AWAIT INSIDE A TEMPLATE EXPRESSION`
-///   ("True if uses runes or top level await or await in template expressions")
-fn detect_await_in_template(ast: &Root, source: &str) -> bool {
-    // Fast path: if the source doesn't contain `await` as a word, bail immediately.
-    if !contains_word(source.as_bytes(), b"await") {
-        return false;
-    }
-
-    fragment_has_template_await(&ast.fragment, source, &ast.arena)
-}
-
-/// Recursively walk a template fragment checking for `{await ...}` ExpressionTags.
-fn fragment_has_template_await(
-    fragment: &crate::ast::template::Fragment,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    for node in &fragment.nodes {
-        if template_node_has_await(node, source, arena) {
-            return true;
-        }
-    }
-    false
 }
 
 /// True if the template fragment contains a real `<slot>` element anywhere
@@ -3286,899 +2551,5 @@ fn node_has_slot_element(node: &crate::ast::template::TemplateNode) -> bool {
                 || b.catch.as_ref().is_some_and(fragment_has_slot_element)
         }
         _ => false,
-    }
-}
-
-/// Check a single template node for `{await ...}` patterns, recursing into children.
-fn template_node_has_await(
-    node: &crate::ast::template::TemplateNode,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::template::TemplateNode;
-
-    match node {
-        // The key check: ExpressionTag with an AwaitExpression.
-        TemplateNode::ExpressionTag(tag) => expression_is_await(&tag.expression, source, arena),
-        // Recurse into element children and attributes
-        TemplateNode::RegularElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&elem.fragment, source, arena)
-        }
-        TemplateNode::Component(comp) => {
-            comp.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&comp.fragment, source, arena)
-        }
-        TemplateNode::IfBlock(block) => {
-            // Also check the `{#if await cond}` test expression — mirrors 2_analyze
-            // which walks `block.test` for has_await.
-            expression_is_await(&block.test, source, arena)
-                || fragment_has_template_await(&block.consequent, source, arena)
-                || block
-                    .alternate
-                    .as_ref()
-                    .map(|alt| fragment_has_template_await(alt, source, arena))
-                    .unwrap_or(false)
-        }
-        TemplateNode::EachBlock(block) => {
-            expression_is_await(&block.expression, source, arena)
-                || fragment_has_template_await(&block.body, source, arena)
-                || block
-                    .fallback
-                    .as_ref()
-                    .map(|fb| fragment_has_template_await(fb, source, arena))
-                    .unwrap_or(false)
-        }
-        TemplateNode::KeyBlock(block) => {
-            expression_is_await(&block.expression, source, arena)
-                || fragment_has_template_await(&block.fragment, source, arena)
-        }
-        // SnippetBlock: official svelte2tsx's `isRunes` sets true for an
-        // AwaitExpression whose ancestor path has no function-expression node —
-        // a SnippetBlock is NOT such a node, so an `await` inside a snippet body
-        // (e.g. `{#snippet}{@const x = await …}{/snippet}`) DOES force runes.
-        // (This is svelte2tsx-specific; the compiler's 2_analyze skips snippets,
-        // but this detector mirrors svelte2tsx, not the compiler.)
-        TemplateNode::SnippetBlock(block) => {
-            fragment_has_template_await(&block.body, source, arena)
-        }
-        // AwaitBlock ({#await expr}) — the `expression` could itself contain an
-        // await (e.g. `{#await await promise}`). Also recurse into the pending /
-        // then / catch sub-fragments since they can contain nested {await ...}
-        // ExpressionTags. Mirrors 2_analyze AwaitBlock fragment_check_features walk.
-        TemplateNode::AwaitBlock(block) => {
-            expression_is_await(&block.expression, source, arena)
-                || block
-                    .pending
-                    .as_ref()
-                    .map(|f| fragment_has_template_await(f, source, arena))
-                    .unwrap_or(false)
-                || block
-                    .then
-                    .as_ref()
-                    .map(|f| fragment_has_template_await(f, source, arena))
-                    .unwrap_or(false)
-                || block
-                    .catch
-                    .as_ref()
-                    .map(|f| fragment_has_template_await(f, source, arena))
-                    .unwrap_or(false)
-        }
-        // SvelteHead, SvelteFragment, SvelteBody, SvelteWindow, SvelteDocument,
-        // SvelteBoundary, SvelteOptions, SvelteSelf — all use the SvelteElement struct.
-        TemplateNode::SvelteHead(elem)
-        | TemplateNode::SvelteFragment(elem)
-        | TemplateNode::SvelteBody(elem)
-        | TemplateNode::SvelteWindow(elem)
-        | TemplateNode::SvelteDocument(elem)
-        | TemplateNode::SvelteBoundary(elem)
-        | TemplateNode::SvelteOptions(elem)
-        | TemplateNode::SvelteSelf(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&elem.fragment, source, arena)
-        }
-        TemplateNode::SvelteComponent(comp) => {
-            comp.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&comp.fragment, source, arena)
-        }
-        TemplateNode::SvelteElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&elem.fragment, source, arena)
-        }
-        TemplateNode::TitleElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&elem.fragment, source, arena)
-        }
-        TemplateNode::SlotElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_await(attr, source, arena))
-                || fragment_has_template_await(&elem.fragment, source, arena)
-        }
-        // HtmlTag ({@html expr}) and RenderTag ({@render expr}) — if the expression
-        // itself is an AwaitExpression (e.g. `{@html await t}`) trigger runes mode.
-        TemplateNode::HtmlTag(tag) => expression_is_await(&tag.expression, source, arena),
-        TemplateNode::RenderTag(tag) => expression_is_await(&tag.expression, source, arena),
-        // `{@const x = await …}` — a top-level await in a const-tag declaration
-        // makes the component async (e.g. inside `<svelte:boundary>`).
-        TemplateNode::ConstTag(ct) => expression_is_await(&ct.declaration, source, arena),
-        // Text, Comment, DeclarationTag, DebugTag, AttachTag — the primary
-        // trigger is ExpressionTag; these are less common.
-        _ => false,
-    }
-}
-
-/// Check if an attribute value contains an await expression in any ExpressionTag part.
-fn attr_has_await(
-    attr: &crate::ast::template::Attribute,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::template::Attribute;
-    use crate::ast::template::AttributeValue;
-    use crate::ast::template::AttributeValuePart;
-
-    // Mirror official's template walk, which sets `isRunes` on ANY top-level
-    // `AwaitExpression` regardless of which attribute/directive it lives in
-    // (e.g. `class:x={await y}`, `style:c={await z}`, `use:a={await b}`,
-    // `bind:v={await w}`). Previously only plain attributes were checked, so
-    // an await confined to a directive failed to flip runes mode and the
-    // `bindings:` field was emitted in legacy (`""`) instead of runes
-    // (`__sveltets_$$bindings('')`) form.
-    let value_has_await = |value: &AttributeValue| match value {
-        AttributeValue::Expression(expr_tag) => {
-            expression_is_await(&expr_tag.expression, source, arena)
-        }
-        AttributeValue::Sequence(parts) => parts.iter().any(|part| {
-            if let AttributeValuePart::ExpressionTag(tag) = part {
-                expression_is_await(&tag.expression, source, arena)
-            } else {
-                false
-            }
-        }),
-        AttributeValue::True(_) => false,
-    };
-    let opt_expr_has_await = |expr: &Option<crate::ast::js::Expression>| {
-        expr.as_ref()
-            .is_some_and(|e| expression_is_await(e, source, arena))
-    };
-
-    match attr {
-        Attribute::Attribute(attr_node) => value_has_await(&attr_node.value),
-        Attribute::SpreadAttribute(s) => expression_is_await(&s.expression, source, arena),
-        Attribute::AttachTag(t) => expression_is_await(&t.expression, source, arena),
-        Attribute::ClassDirective(d) => expression_is_await(&d.expression, source, arena),
-        Attribute::BindDirective(d) => expression_is_await(&d.expression, source, arena),
-        Attribute::StyleDirective(d) => value_has_await(&d.value),
-        Attribute::OnDirective(d) => opt_expr_has_await(&d.expression),
-        Attribute::TransitionDirective(d) => opt_expr_has_await(&d.expression),
-        Attribute::AnimateDirective(d) => opt_expr_has_await(&d.expression),
-        Attribute::UseDirective(d) => opt_expr_has_await(&d.expression),
-        Attribute::LetDirective(_) => false,
-    }
-}
-
-/// Check if an Expression node is (or begins with) an AwaitExpression.
-///
-/// For `Typed` expressions, checks the top-level JsNode variant.
-/// For `Lazy` expressions (source spans), checks the source text.
-/// For `Value` (JSON) expressions, checks the JSON `type` field.
-fn expression_is_await(
-    expr: &crate::ast::js::Expression,
-    source: &str,
-    _arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::js::Expression;
-    use crate::ast::typed_expr::JsNode;
-
-    // A top-level `await` ANYWHERE inside the expression makes the component
-    // async (→ runes), not only when the whole expression IS an await — e.g.
-    // `{(await user).name}`, `{foo(await x)}`, `{cond ? await a : b}`. Fast-path
-    // the direct `{await x}` form, then scan the expression's source span for
-    // `await` as a word, which covers every nesting depth. (A literal "await"
-    // string is a rare false positive; svelte itself treats such components as
-    // async too once a template `await` is present.)
-    let direct = match expr {
-        Expression::Typed(te) => matches!(&te.node, JsNode::AwaitExpression { .. }),
-        Expression::Lazy { .. } => false,
-    };
-    if direct {
-        return true;
-    }
-    // Non-direct: a `await` nested in e.g. `(await user).name` / `foo(await x)`
-    // still makes the component async — BUT an `await` inside a nested function
-    // (`() => await x`) is a different scope and must NOT count (mirrors the
-    // upstream `scope === rootScope` rule). Approximate the function-boundary
-    // check on the source span: count the `await` only when the expression
-    // contains no function boundary (`=>` / `function`), which keeps the common
-    // member/call/conditional cases without over-triggering on callbacks.
-    if let (Some(s), Some(e)) = (expr.start(), expr.end()) {
-        let (s, e) = (s as usize, e as usize);
-        if s < e && e <= source.len() {
-            let span = &source.as_bytes()[s..e];
-            return contains_word(span, b"await")
-                && !span.windows(2).any(|w| w == b"=>")
-                && !contains_word(span, b"function");
-        }
-    }
-    false
-}
-
-// =============================================================================
-// Rune-global-in-template detection
-//
-// Mirrors the official `checkGlobalsForRunes` pass which treats every undeclared
-// `$state` / `$derived` / `$effect` identifier anywhere in the component (script
-// OR template) as evidence of runes mode.  The instance-script scanner handles
-// the `<script>` side; these helpers cover the template side so that components
-// with NO `<script>` but with e.g. `{$state.eager(x)}` are correctly classified.
-//
-// Reference: language-tools/packages/svelte2tsx/src/svelte2tsx/index.ts
-//   `exportedNames.checkGlobalsForRunes(implicitStoreValues.getGlobals())`
-// Reference: language-tools/packages/svelte2tsx/src/svelte2tsx/nodes/ExportedNames.ts
-//   `hasRunesGlobals = isSvelte5Plus && globals.some(g => ['$state','$derived','$effect'].includes(g))`
-// =============================================================================
-
-/// Detect any `$state`/`$derived`/`$effect` rune-global reference inside the
-/// template fragment.
-///
-/// Fast-path: returns `false` immediately when none of the three magic words
-/// appear as a word boundary in the raw source.  The AST walk is only done when
-/// a quick substring match succeeds.
-fn detect_rune_global_in_template(
-    ast: &Root,
-    source: &str,
-    instance_value_names: &std::collections::HashSet<String>,
-) -> bool {
-    // Fast path: if neither $state, $derived, nor $effect appears in the source
-    // as a word start, bail immediately.  These identifiers always start with `$`
-    // so a simple substring check is conservative (won't false-positive on
-    // e.g. `some_$state_like_string` since we still walk the AST after this).
-    if !source.contains("$state") && !source.contains("$derived") && !source.contains("$effect") {
-        return false;
-    }
-
-    // Populate the shadowed-rune set: a `state`/`derived`/`effect` declared as an
-    // instance variable makes the matching `$`-rune a store sub, not a rune.
-    SHADOWED_RUNE_BASES.with(|s| {
-        let mut set = s.borrow_mut();
-        set.clear();
-        for base in ["state", "derived", "effect"] {
-            if instance_value_names.contains(base) {
-                set.insert(base.to_string());
-            }
-        }
-    });
-    let result = fragment_has_template_rune_global(&ast.fragment, source, &ast.arena);
-    SHADOWED_RUNE_BASES.with(|s| s.borrow_mut().clear());
-    result
-}
-
-/// Recursively walk a template fragment checking for rune-global references.
-fn fragment_has_template_rune_global(
-    fragment: &crate::ast::template::Fragment,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    for node in &fragment.nodes {
-        if template_node_has_rune_global(node, source, arena) {
-            return true;
-        }
-    }
-    false
-}
-
-/// Check a single template node for rune-global references, recursing into children.
-fn template_node_has_rune_global(
-    node: &crate::ast::template::TemplateNode,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::template::TemplateNode;
-
-    match node {
-        // The primary check: ExpressionTag { expr } — check if the expression
-        // references a $state/$derived/$effect global.
-        TemplateNode::ExpressionTag(tag) => {
-            expression_references_rune_global(&tag.expression, source, arena)
-        }
-        // Recurse into element children and attributes
-        TemplateNode::RegularElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&elem.fragment, source, arena)
-        }
-        TemplateNode::Component(comp) => {
-            comp.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&comp.fragment, source, arena)
-        }
-        TemplateNode::IfBlock(block) => {
-            expression_references_rune_global(&block.test, source, arena)
-                || fragment_has_template_rune_global(&block.consequent, source, arena)
-                || block
-                    .alternate
-                    .as_ref()
-                    .map(|alt| fragment_has_template_rune_global(alt, source, arena))
-                    .unwrap_or(false)
-        }
-        TemplateNode::EachBlock(block) => {
-            expression_references_rune_global(&block.expression, source, arena)
-                || fragment_has_template_rune_global(&block.body, source, arena)
-                || block
-                    .fallback
-                    .as_ref()
-                    .map(|fb| fragment_has_template_rune_global(fb, source, arena))
-                    .unwrap_or(false)
-        }
-        TemplateNode::KeyBlock(block) => {
-            expression_references_rune_global(&block.expression, source, arena)
-                || fragment_has_template_rune_global(&block.fragment, source, arena)
-        }
-        // SnippetBlock: official's global collection (checkGlobalsForRunes via
-        // implicitStoreValues) walks the whole component including snippet
-        // bodies, so a rune call inside a snippet (`{#snippet}{@const x =
-        // $derived(…)}{/snippet}`) forces runes mode. Recurse into the body.
-        TemplateNode::SnippetBlock(block) => {
-            fragment_has_template_rune_global(&block.body, source, arena)
-        }
-        TemplateNode::AwaitBlock(block) => {
-            expression_references_rune_global(&block.expression, source, arena)
-                || block
-                    .pending
-                    .as_ref()
-                    .map(|f| fragment_has_template_rune_global(f, source, arena))
-                    .unwrap_or(false)
-                || block
-                    .then
-                    .as_ref()
-                    .map(|f| fragment_has_template_rune_global(f, source, arena))
-                    .unwrap_or(false)
-                || block
-                    .catch
-                    .as_ref()
-                    .map(|f| fragment_has_template_rune_global(f, source, arena))
-                    .unwrap_or(false)
-        }
-        // SvelteHead, SvelteFragment, SvelteBody, SvelteWindow, SvelteDocument,
-        // SvelteBoundary, SvelteOptions, SvelteSelf — all use the SvelteElement struct.
-        TemplateNode::SvelteHead(elem)
-        | TemplateNode::SvelteFragment(elem)
-        | TemplateNode::SvelteBody(elem)
-        | TemplateNode::SvelteWindow(elem)
-        | TemplateNode::SvelteDocument(elem)
-        | TemplateNode::SvelteBoundary(elem)
-        | TemplateNode::SvelteOptions(elem)
-        | TemplateNode::SvelteSelf(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&elem.fragment, source, arena)
-        }
-        TemplateNode::SvelteComponent(comp) => {
-            comp.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&comp.fragment, source, arena)
-        }
-        TemplateNode::SvelteElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&elem.fragment, source, arena)
-        }
-        TemplateNode::TitleElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&elem.fragment, source, arena)
-        }
-        TemplateNode::SlotElement(elem) => {
-            elem.attributes
-                .iter()
-                .any(|attr| attr_has_rune_global(attr, source, arena))
-                || fragment_has_template_rune_global(&elem.fragment, source, arena)
-        }
-        // HtmlTag ({@html expr}) and RenderTag ({@render expr})
-        TemplateNode::HtmlTag(tag) => {
-            expression_references_rune_global(&tag.expression, source, arena)
-        }
-        TemplateNode::RenderTag(tag) => {
-            expression_references_rune_global(&tag.expression, source, arena)
-        }
-        // AttachTag ({@attach expr}) — the expression may contain nested
-        // rune calls, e.g. `{@attach $effect(() => { ... })}`.
-        // Reference: official svelte2tsx collects `@attach` expression globals
-        // via `implicitStoreValues` just like any other template expression.
-        TemplateNode::AttachTag(tag) => {
-            expression_references_rune_global(&tag.expression, source, arena)
-        }
-        // `{@const x = $derived(…)}` and `{let x = $state(0), …}` carry rune
-        // calls in their declaration; official collects their globals like any
-        // other template expression, so a runes-only component with no script
-        // (only template declaration tags) still enters runes mode. The
-        // declaration is a `VariableDeclaration` (which the typed/JSON rune
-        // walkers don't descend into), so scan the tag's source slice directly.
-        TemplateNode::ConstTag(tag) => {
-            let (s, e) = (tag.start as usize, tag.end as usize);
-            s < e && e <= source.len() && lazy_slice_references_rune_global(&source[s..e])
-        }
-        TemplateNode::DeclarationTag(tag) => {
-            let (s, e) = (tag.start as usize, tag.end as usize);
-            s < e && e <= source.len() && lazy_slice_references_rune_global(&source[s..e])
-        }
-        _ => false,
-    }
-}
-
-/// Check if an attribute (of any kind) contains a rune-global reference.
-///
-/// Covers all `Attribute` variants:
-/// - `Attribute` (plain attribute with expression/sequence value)
-/// - `SpreadAttribute` (spread expression)
-/// - `AttachTag` (`{@attach expr}` used inside element attribute position)
-/// - All directives: `bind:`, `on:`, `class:`, `style:`, `transition:`,
-///   `animate:`, `use:`, `let:` — each may carry an expression value.
-///
-/// Reference: official svelte2tsx passes ALL template expressions through
-/// `implicitStoreValues` (which collects globals), not just plain attributes.
-/// Mirrors the comprehensive directive coverage in `attr_has_await`.
-fn attr_has_rune_global(
-    attr: &crate::ast::template::Attribute,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::template::Attribute;
-    use crate::ast::template::AttributeValue;
-    use crate::ast::template::AttributeValuePart;
-
-    match attr {
-        // Plain attribute: check expression / sequence values.
-        Attribute::Attribute(attr_node) => match &attr_node.value {
-            AttributeValue::Expression(expr_tag) => {
-                expression_references_rune_global(&expr_tag.expression, source, arena)
-            }
-            AttributeValue::Sequence(parts) => parts.iter().any(|part| {
-                if let AttributeValuePart::ExpressionTag(tag) = part {
-                    expression_references_rune_global(&tag.expression, source, arena)
-                } else {
-                    false
-                }
-            }),
-            AttributeValue::True(_) => false,
-        },
-
-        // Spread attribute: `{...expr}` — check the spread expression.
-        Attribute::SpreadAttribute(spread) => {
-            expression_references_rune_global(&spread.expression, source, arena)
-        }
-
-        // AttachTag in attribute position: `{@attach expr}`.
-        Attribute::AttachTag(attach) => {
-            expression_references_rune_global(&attach.expression, source, arena)
-        }
-
-        // bind:name={expr} — expression is always present.
-        Attribute::BindDirective(bind) => {
-            expression_references_rune_global(&bind.expression, source, arena)
-        }
-
-        // on:event={handler} — expression is Optional<Expression>.
-        Attribute::OnDirective(on) => on
-            .expression
-            .as_ref()
-            .is_some_and(|e| expression_references_rune_global(e, source, arena)),
-
-        // class:name={expr} — expression is always present.
-        Attribute::ClassDirective(class) => {
-            expression_references_rune_global(&class.expression, source, arena)
-        }
-
-        // style:property={value} — value is AttributeValue (same shape as plain attr).
-        Attribute::StyleDirective(style) => match &style.value {
-            AttributeValue::Expression(expr_tag) => {
-                expression_references_rune_global(&expr_tag.expression, source, arena)
-            }
-            AttributeValue::Sequence(parts) => parts.iter().any(|part| {
-                if let AttributeValuePart::ExpressionTag(tag) = part {
-                    expression_references_rune_global(&tag.expression, source, arena)
-                } else {
-                    false
-                }
-            }),
-            AttributeValue::True(_) => false,
-        },
-
-        // transition:name={params} / in: / out: — expression is Optional.
-        Attribute::TransitionDirective(t) => t
-            .expression
-            .as_ref()
-            .is_some_and(|e| expression_references_rune_global(e, source, arena)),
-
-        // animate:name={params} — expression is Optional.
-        Attribute::AnimateDirective(a) => a
-            .expression
-            .as_ref()
-            .is_some_and(|e| expression_references_rune_global(e, source, arena)),
-
-        // use:action={params} — expression is Optional.
-        Attribute::UseDirective(u) => u
-            .expression
-            .as_ref()
-            .is_some_and(|e| expression_references_rune_global(e, source, arena)),
-
-        // let:item — rarely carries a rune call, but check for completeness.
-        Attribute::LetDirective(l) => l
-            .expression
-            .as_ref()
-            .is_some_and(|e| expression_references_rune_global(e, source, arena)),
-    }
-}
-
-// Rune base names (`state`/`derived`/`effect`) that are SHADOWED by an
-// instance-script variable of the same name. Official `getGlobals()` removes
-// declared variables, so `$state.snapshot(state)` where `state` is declared
-// is a store auto-subscription, NOT a `$state` rune — the component stays
-// legacy. Set (read-only) for the duration of `detect_rune_global_in_template`
-// on this thread and cleared right after; never accumulated across calls.
-thread_local! {
-    static SHADOWED_RUNE_BASES: std::cell::RefCell<std::collections::HashSet<String>> =
-        std::cell::RefCell::new(std::collections::HashSet::new());
-}
-
-fn rune_base_is_shadowed(base: &str) -> bool {
-    SHADOWED_RUNE_BASES.with(|s| s.borrow().contains(base))
-}
-
-/// Returns `true` if `name` is one of the three rune-global identifiers and is
-/// not shadowed by a declared instance variable.
-#[inline]
-fn is_rune_global_name(name: &str) -> bool {
-    matches!(name, "$state" | "$derived" | "$effect") && !rune_base_is_shadowed(&name[1..])
-}
-
-/// Check whether an Expression node references a `$state`/`$derived`/`$effect` global.
-///
-/// For `Typed` expressions, walks the JsNode tree stored in the parse arena.
-/// For `Lazy` expressions (raw source spans), scans the source text.
-/// For `Value` (JSON) expressions, inspects the JSON AST.
-///
-/// The walk is deliberately shallow-but-sufficient: it recurses into the callee
-/// of a CallExpression and the object of a MemberExpression (the two patterns
-/// that can reference a rune global — `$state(x)` and `$state.eager(x)`) but
-/// does NOT recurse into every sub-expression.  Template expressions that use
-/// rune globals almost always have the global as the outermost callee or
-/// member-expression object, so this covers all real-world cases while keeping
-/// the implementation simple and fast.
-///
-/// Reference: ExportedNames.ts `checkGlobalsForRunes` which sets
-///   `hasRunesGlobals` when any of `$state`/`$derived`/`$effect` appear as an
-///   undeclared identifier anywhere in the component globals set.
-fn expression_references_rune_global(
-    expr: &crate::ast::js::Expression,
-    source: &str,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::js::Expression;
-
-    match expr {
-        Expression::Typed(te) => js_node_references_rune_global(&te.node, arena),
-        Expression::Lazy { start, end, .. } => {
-            // Raw source slice — scan for `$state`, `$derived`, `$effect` as
-            // identifier-like occurrences.  We already know the full source
-            // contains one of these strings (fast-path in
-            // `detect_rune_global_in_template`), so this walk is uncommon.
-            let s = *start as usize;
-            let e = *end as usize;
-            if s < e && e <= source.len() {
-                let slice = &source[s..e];
-                lazy_slice_references_rune_global(slice)
-            } else {
-                false
-            }
-        }
-    }
-}
-
-/// Check whether a callee `JsNode` directly IS a rune-global call target.
-///
-/// A callee is a rune-global target when it is:
-///   - An `Identifier` named `$state`/`$derived`/`$effect`  (direct call: `$state(x)`)
-///   - A `MemberExpression` whose object is such an identifier  (`$state.eager(x)`)
-///
-/// This intentionally does NOT recurse further — if the callee is something more
-/// complex, it is not a rune call pattern.
-#[inline]
-fn js_callee_is_rune_global(
-    callee: &crate::ast::typed_expr::JsNode,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::typed_expr::JsNode;
-    match callee {
-        JsNode::Identifier { name, .. } => is_rune_global_name(name.as_str()),
-        JsNode::MemberExpression { object, .. } => {
-            let obj = arena.get_js_node(*object);
-            matches!(obj, JsNode::Identifier { name, .. } if is_rune_global_name(name.as_str()))
-        }
-        _ => false,
-    }
-}
-
-/// Walk a `JsNode` (typed AST node stored in the parse arena) looking for a
-/// `$state`/`$derived`/`$effect` rune call anywhere in the expression tree.
-///
-/// A RUNE CALL means the global is used as a call callee or as the object of a
-/// member-expression that is itself used as a call callee.  A bare `$state`
-/// identifier that is just a store auto-subscription (`{$state}`) does NOT match.
-///
-/// Handles patterns like:
-///   - `$state(x)`                     → CallExpression callee = Identifier "$state"
-///   - `$state.eager(x)`               → CallExpression callee = MemberExpression { object = Identifier "$state" }
-///   - `$effect.pre(() => …)`          → same
-///   - `foo($state(x))`                → arguments contain a rune CallExpression
-///   - `a === '/' ? $state(x) : null`  → ConditionalExpression branches
-///   - `() => $effect(() => {})`       → ArrowFunctionExpression body
-///   - `{@attach $effect(() => {})}`   → ArrowFunctionExpression body in AttachTag
-///   - `[..., $state(x)]`              → ArrayExpression element
-///   - `{ k: $derived(v) }`            → ObjectExpression property value
-///
-/// Does NOT match:
-///   - `{$state}` (bare store auto-subscription; no call)
-///   - `$state + 1` (store ref in arithmetic; no call)
-///
-/// Reference: official `implicitStoreValues` collects ALL undeclared globals,
-/// including those inside nested function bodies passed to directives.
-fn js_node_references_rune_global(
-    node: &crate::ast::typed_expr::JsNode,
-    arena: &crate::ast::arena::ParseArena,
-) -> bool {
-    use crate::ast::typed_expr::JsNode;
-    match node {
-        // CallExpression: the callee must be a rune-global target (direct call
-        // `$state(...)` or member-call `$state.eager(...)`).  Also recurse into
-        // arguments so nested rune calls like `foo($state(x))` are caught.
-        JsNode::CallExpression {
-            callee, arguments, ..
-        } => {
-            let callee_node = arena.get_js_node(*callee);
-            if js_callee_is_rune_global(callee_node, arena) {
-                return true;
-            }
-            // Recurse into arguments to catch `foo($state(x))`.
-            let args = arena.get_js_children(*arguments);
-            args.iter()
-                .any(|arg| js_node_references_rune_global(arg, arena))
-        }
-
-        // ConditionalExpression: check test, consequent, alternate.
-        // E.g. `$state.eager(x) === '/' ? 'page' : null` — the test is the
-        // BinaryExpression; we recurse into it and then into the call.
-        JsNode::ConditionalExpression {
-            test,
-            consequent,
-            alternate,
-            ..
-        } => {
-            js_node_references_rune_global(arena.get_js_node(*test), arena)
-                || js_node_references_rune_global(arena.get_js_node(*consequent), arena)
-                || js_node_references_rune_global(arena.get_js_node(*alternate), arena)
-        }
-
-        // BinaryExpression / LogicalExpression: check both sides.
-        // E.g. `$state.eager(pathname) === '/'` — the left side is the call.
-        JsNode::BinaryExpression { left, right, .. }
-        | JsNode::LogicalExpression { left, right, .. } => {
-            js_node_references_rune_global(arena.get_js_node(*left), arena)
-                || js_node_references_rune_global(arena.get_js_node(*right), arena)
-        }
-
-        // ArrowFunctionExpression: recurse into the body.
-        // Covers `{@attach $effect(() => { ... })}` and
-        // `use:action={() => $state(x)}` patterns.
-        // The body is a JsNodeId pointing to either a BlockStatement or an
-        // expression (when `expression: true`).
-        JsNode::ArrowFunctionExpression { body, .. } => {
-            let body_node = arena.get_js_node(*body);
-            js_node_references_rune_global(body_node, arena)
-        }
-
-        // FunctionExpression: recurse into the body (a BlockStatement or None).
-        // E.g. `use:action={function() { $effect(() => {}); }}`.
-        JsNode::FunctionExpression { body, .. } => body
-            .map(|b| {
-                let body_node = arena.get_js_node(b);
-                js_node_references_rune_global(body_node, arena)
-            })
-            .unwrap_or(false),
-
-        // BlockStatement: recurse into each statement.
-        // Reached from FunctionExpression / ArrowFunctionExpression bodies.
-        JsNode::BlockStatement { body, .. } => {
-            let stmts = arena.get_js_children(*body);
-            stmts
-                .iter()
-                .any(|s| js_node_references_rune_global(s, arena))
-        }
-
-        // ExpressionStatement: unwrap to the inner expression.
-        JsNode::ExpressionStatement { expression, .. } => {
-            js_node_references_rune_global(arena.get_js_node(*expression), arena)
-        }
-
-        // ObjectExpression: recurse into property values.
-        // E.g. `use:action={{ key: $state(x) }}`.
-        JsNode::ObjectExpression { properties, .. } => {
-            let props = arena.get_js_children(*properties);
-            props.iter().any(|p| {
-                if let JsNode::Property { value, .. } = p {
-                    js_node_references_rune_global(arena.get_js_node(*value), arena)
-                } else {
-                    false
-                }
-            })
-        }
-
-        // ArrayExpression: recurse into elements (elements are inline, not arena-indexed).
-        // E.g. `{[$state(a), $derived(b)]}`.
-        JsNode::ArrayExpression { elements, .. } => elements.iter().any(|elem| {
-            elem.as_ref()
-                .is_some_and(|e| js_node_references_rune_global(e, arena))
-        }),
-
-        // SequenceExpression: recurse into each sub-expression.
-        // E.g. `{(doSomething(), $state(x))}`.
-        JsNode::SequenceExpression { expressions, .. } => {
-            let exprs = arena.get_js_children(*expressions);
-            exprs
-                .iter()
-                .any(|e| js_node_references_rune_global(e, arena))
-        }
-
-        // AwaitExpression: recurse into the argument.
-        // Rare in template context but possible.
-        JsNode::AwaitExpression { argument, .. } => {
-            js_node_references_rune_global(arena.get_js_node(*argument), arena)
-        }
-
-        // UnaryExpression: recurse into argument (e.g. `!$state(x)`).
-        JsNode::UnaryExpression { argument, .. } => {
-            js_node_references_rune_global(arena.get_js_node(*argument), arena)
-        }
-
-        // AssignmentExpression: check right-hand side.
-        // E.g. `x = $state(0)` inside a function body.
-        JsNode::AssignmentExpression { right, .. } => {
-            js_node_references_rune_global(arena.get_js_node(*right), arena)
-        }
-
-        // VariableDeclaration / VariableDeclarator: recurse into each
-        // declarator's initializer — e.g. `const state = $state({…})` inside an
-        // event-handler arrow body (`onsubmit={e => { const s = $state(…) }}`).
-        JsNode::VariableDeclaration { declarations, .. } => {
-            let decls = arena.get_js_children(*declarations);
-            decls
-                .iter()
-                .any(|d| js_node_references_rune_global(d, arena))
-        }
-        JsNode::VariableDeclarator { init, .. } => init
-            .map(|i| js_node_references_rune_global(arena.get_js_node(i), arena))
-            .unwrap_or(false),
-
-        // ReturnStatement / IfStatement bodies can also host rune calls.
-        JsNode::ReturnStatement { argument, .. } => argument
-            .map(|a| js_node_references_rune_global(arena.get_js_node(a), arena))
-            .unwrap_or(false),
-
-        // Bare Identifier (e.g. `{$state}` — store auto-subscription) → NOT a rune call.
-        // MemberExpression without being called (e.g. `$state.value` as a bare expr) → NOT a rune call.
-        // These are legitimate store/object references, not rune invocations.
-        _ => false,
-    }
-}
-
-/// Scan a raw source slice (from a `Lazy` expression) for a rune-global CALL.
-///
-/// Only triggers when `$state`/`$derived`/`$effect` is immediately followed by
-/// `(` (direct call) or `.` (member call like `$state.eager(…)`).  A bare
-/// `$state` with no following `(` or `.` is a store auto-subscription reference
-/// and must NOT trigger runes mode.
-fn lazy_slice_references_rune_global(slice: &str) -> bool {
-    for candidate in &["$state", "$derived", "$effect"] {
-        // A rune whose base is shadowed by a declared instance var is a store
-        // sub, not a rune (see SHADOWED_RUNE_BASES).
-        if rune_base_is_shadowed(&candidate[1..]) {
-            continue;
-        }
-        let mut search_from = 0;
-        while let Some(rel) = slice[search_from..].find(candidate) {
-            let idx = search_from + rel;
-            let after = idx + candidate.len();
-            if after < slice.len() {
-                let next = slice.as_bytes()[after];
-                // Require `(` (direct call) or `.` (member call).
-                // Also ensure the match is not inside a longer identifier
-                // (e.g. `$state_machine` — `$` is a valid JS identifier char).
-                if next == b'(' || next == b'.' {
-                    return true;
-                }
-            }
-            search_from = idx + 1;
-        }
-    }
-    false
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashSet;
-
-    fn set(names: &[&str]) -> HashSet<String> {
-        names.iter().map(|s| s.to_string()).collect()
-    }
-
-    #[test]
-    fn typeof_local_value_blocks_hoist_but_import_does_not() {
-        // `typeof Button` where Button is imported → hoistable (false).
-        let imports = set(&["Button"]);
-        assert!(!type_text_typeof_references_local_value(
-            "ComponentProps<typeof Button>",
-            &set(&[]),
-            &imports,
-            &set(&[]),
-        ));
-        // `typeof localVal` where localVal is an instance-script value → block (true).
-        assert!(type_text_typeof_references_local_value(
-            "typeof localVal",
-            &set(&["localVal"]),
-            &set(&[]),
-            &set(&[]),
-        ));
-        // No `typeof` at all → false.
-        assert!(!type_text_typeof_references_local_value(
-            "{ a: string; b: number }",
-            &set(&["localVal"]),
-            &set(&[]),
-            &set(&[]),
-        ));
-        // A module-scope import via `typeof` is also hoistable (false).
-        assert!(!type_text_typeof_references_local_value(
-            "typeof ModuleThing",
-            &set(&[]),
-            &set(&[]),
-            &set(&["ModuleThing"]),
-        ));
-    }
-
-    #[test]
-    fn test_derive_component_name() {
-        // Ground-truth cases verified against the official svelte2tsx classNameFromFilename.
-        assert_eq!(derive_component_name("App.svelte"), "App");
-        assert_eq!(derive_component_name("my-component.svelte"), "MyComponent");
-        assert_eq!(derive_component_name("my_component.svelte"), "MyComponent");
-        assert_eq!(derive_component_name("path/to/Input.svelte"), "Input");
-        assert_eq!(derive_component_name("123.svelte"), "A3");
-        assert_eq!(derive_component_name("1.svelte"), "A1");
-        assert_eq!(derive_component_name("foo.bar.svelte"), "Foo");
-        assert_eq!(derive_component_name("ABCWidget.svelte"), "ABCWidget");
-        assert_eq!(derive_component_name("XMLHttp.svelte"), "XMLHttp");
-        assert_eq!(derive_component_name("a1b2.svelte"), "A1b2");
-        assert_eq!(derive_component_name("_x.svelte"), "X");
-        assert_eq!(derive_component_name("two words.svelte"), "Twowords");
-        assert_eq!(derive_component_name(".svelte"), "A");
     }
 }
